+++ pkg/scheme/product.go
@@ -1,7 +1,7 @@
 package scheme
 
 type Product struct {
-	ID    int32   `json:"id"`
+	ID    int     `json:"id"`
 	Name  string  `json:"name"`
-	Price float32 `json:"price"`
+	Price float64 `json:"price"`
 }
